scripts/etl/requirements.txt
pymongo>=4.0.0
psycopg2-binary>=2.9.0
pandas>=1.5.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0

